feat(Context): swap to TS
